Change Puzzle League chaining a bit
Chains in general now only use a flag instead of a counter, because the
original game only has a counter per-playfield instead of per-tile. Also
when chain flags are set, ALL tiles above are set, even if there are
gaps.

I've also started working on including an alternate implementation of
the Puzzle League physics but they're not incorporated yet.